revert: roll back webview native-owns-tokens auth (#223)
PR #223(웹/앱 인증 이원화 — AuthStrategy seam + 네이티브-소유-토큰)을 되돌려 iron-session 모델로 롤백.
- AuthStrategy seam(strategies/*) · webviewStrategy · replyBridge 제거,
  AuthContext·tokenStore 를 #223 이전(쿠키 기반) 상태로 복원.
- route.ts 의 POST /api/session 익스체인지(iron-session 웹뷰 진입)는 그대로 유지.
- #224(getMe) · #222(졸업 UI) 는 파일 비중첩이라 영향 없음.
대신 패턴 3(세션 전송용 1회용 토큰, Auth0 Native-to-Web SSO 식) 검토 예정.

This reverts merge commit 257e77a (PR #223).
