+++ src/lib/webview/index.ts
@@ -1,8 +1,1 @@
 export { isInWebView, postBridgeMessage, navigateNative, navigateBack, redirectToHome, withdraw } from './bridge';
-export {
-  hasNativeAuthCapability,
-  readNativeAuth,
-  requestNativeRefresh,
-  signalNativeLogout,
-} from './replyBridge';
-export type { NativeAuthPayload } from './replyBridge';
